requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# NLP Libraries
spacy>=3.7.0
transformers>=4.35.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web Scraping (for data collection)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Utilities
pyyaml>=6.0.1
python-multipart>=0.0.6

# Optional: Telegram (for data collection)
python-telegram-bot>=20.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
